reorg domain-infra
