+++ blockchain-watcher/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@wormhole-foundation/event-watcher",
+  "name": "@wormhole-foundation/blockchain-watcher",
   "version": "0.0.0",
   "description": "A process for watching blockchain events and moving them to persistent storage",
   "main": "index.js",
   "scripts": {
     "start": "node lib/index.js",
     "build": "tsc",
-    "dev": "USE_ENV_FILE=true ts-node src/index2.ts",
+    "dev": "USE_ENV_FILE=true ts-node src/index.ts",
     "prettier": "prettier --write ."
   },
   "author": "chase-45",
