+++ blockchain-watcher/src/infrastructure/handlers/AbstractHandler.ts
