+++ blockchain-watcher/src/infrastructure/handlers/sendEventHandler.ts
