+++ blockchain-watcher/src/infrastructure/utils/websocket.ts
