+++ blockchain-watcher/src/infrastructure/watchers/AbstractWatcher.ts
