+++ blockchain-watcher/test/domain/job.test.ts
@@ -1,0 +1,7 @@
+import { describe, it, expect } from "@jest/globals";
+
+describe("Job", () => {
+  it("should be able to validate a job", () => {
+    expect(true).toBe(true);
+  });
+});
